Move compatibility workarounds for QMutex into new subdirectory
