+++ src/engine/controls/cuecontrol.h
@@ -11,8 +11,8 @@
 #include "preferences/usersettings.h"
 #include "track/cue.h"
 #include "track/track_decl.h"
+#include "util/compatibility/qmutex.h"
 #include "util/parented_ptr.h"
-#include "util/qtmutex.h"
 
 #define NUM_HOT_CUES 37
