+++ src/util/mutex.h
@@ -8,7 +8,7 @@
 #include <QMutexLocker>
 #include <QReadWriteLock>
 
-#include "util/qtmutex.h"
+#include "util/compatibility/qmutex.h"
 #include "util/thread_annotations.h"
 
 class CAPABILITY("mutex") MMutex {
