+++ src/util/statsmanager.cpp
@@ -9,7 +9,7 @@
 #include "moc_statsmanager.cpp"
 #include "util/cmdlineargs.h"
 #include "util/compatibility.h"
-#include "util/qtmutex.h"
+#include "util/compatibility/qmutex.h"
 
 // In practice we process stats pipes about once a minute @1ms latency.
 const int kStatsPipeSize = 1 << 10;
