+++ src/waveform/waveform.h
@@ -10,7 +10,7 @@
 
 #include "util/class.h"
 #include "util/compatibility.h"
-#include "util/qtmutex.h"
+#include "util/compatibility/qmutex.h"
 
 enum FilterIndex { Low = 0, Mid = 1, High = 2, FilterCount = 3};
 enum ChannelIndex { Left = 0, Right = 1, ChannelCount = 2};
@@ -32,7 +32,7 @@ class Waveform {
   public:
     enum class SaveState {
         NotSaved = 0,
-        SavePending, 
+        SavePending,
         Saved
     };
 
@@ -84,7 +84,7 @@ class Waveform {
         return m_saveState;
     }
 
-    // AnalysisDAO needs to be able to change the state to savePending when finished 
+    // AnalysisDAO needs to be able to change the state to savePending when finished
     // so we mark this as const and m_saveState mutable.
     void setSaveState(SaveState eState) const {
         m_saveState = eState;
