+++ src/widget/wlibrary.h
@@ -8,7 +8,7 @@
 
 #include "library/libraryview.h"
 #include "skin/legacy/skincontext.h"
-#include "util/qtmutex.h"
+#include "util/compatibility/qmutex.h"
 #include "widget/wbasewidget.h"
 
 class KeyboardEventFilter;
